+++ saleor/core/management/commands/populatedb.py
@@ -1,7 +1,6 @@
 from io import StringIO
 
 from django.apps import apps
-from django.conf import settings
 from django.core.management import call_command
 from django.core.management.base import BaseCommand
 from django.db import connection
@@ -67,10 +66,6 @@ class Command(BaseCommand):
             cursor.execute("PRAGMA temp_store = MEMORY;")
             cursor.execute("PRAGMA synchronous = OFF;")
 
-    def populate_search_index(self):
-        if settings.ES_URL:
-            call_command("search_index", "--rebuild", force=True)
-
     def sequence_reset(self):
         """Run a SQL sequence reset on all saleor.* apps.
 
@@ -116,7 +111,5 @@ class Command(BaseCommand):
             msg = create_superuser(credentials)
             self.stdout.write(msg)
             add_address_to_admin(credentials["email"])
-        if not options["withoutsearch"]:
-            self.populate_search_index()
         if not options["skipsequencereset"]:
             self.sequence_reset()