+++ saleor/page/models.py
@@ -1,5 +1,4 @@
 from django.db import models
-from django.urls import reverse
 from django.utils.translation import pgettext_lazy
 from draftjs_sanitizer import clean_draft_js
 
@@ -36,8 +35,10 @@ class Page(SeoModel, PublishableModel):
     def __str__(self):
         return self.title
 
-    def get_absolute_url(self):
-        return reverse("page:details", kwargs={"slug": self.slug})
+    # Deprecated. To remove in #5022
+    @staticmethod
+    def get_absolute_url():
+        return ""
 
 
 class PageTranslation(SeoModelTranslation):