+++ saleor/payment/gateways/razorpay/plugin.py
@@ -129,10 +129,6 @@ class RazorpayGatewayPlugin(BasePlugin):
         return process_payment(payment_information, self._get_gateway_config())
 
     @require_active_plugin
-    def get_payment_template(self, previous_value) -> str:
-        return self._get_gateway_config().template_path
-
-    @require_active_plugin
     def create_form(
         self, data, payment_information: "PaymentData", previous_value
     ) -> "forms.Form":