+++ tests/test_core.py
@@ -4,24 +4,20 @@ from unittest.mock import Mock, patch
 from urllib.parse import urljoin
 
 import pytest
-from django.shortcuts import reverse
 from django.templatetags.static import static
-from django.test import Client, RequestFactory, override_settings
-from django.urls import translate_url
+from django.test import RequestFactory, override_settings
 from measurement.measures import Weight
-from prices import Money
 
 from saleor.account.models import Address, User
 from saleor.account.utils import create_superuser
 from saleor.core.storages import S3MediaStorage
+from saleor.core.templatetags.placeholder import placeholder
 from saleor.core.utils import (
     Country,
     build_absolute_uri,
     create_thumbnails,
-    format_money,
     get_client_ip,
     get_country_by_ip,
-    get_country_name_by_code,
     get_currency_for_country,
     random_data,
 )
@@ -44,11 +40,6 @@ type_schema = {
         "is_shipping_required": True,
     }
 }
-
-
-def test_format_money():
-    money = Money("123.99", "USD")
-    assert format_money(money) == "$123.99"
 
 
 @pytest.mark.parametrize(
@@ -109,13 +100,6 @@ def test_create_superuser(db, client, media_root):
     # Test duplicating
     create_superuser(credentials)
     assert User.objects.all().count() == 1
-    # Test logging in
-    response = client.post(
-        reverse("account:login"),
-        {"username": credentials["email"], "password": credentials["password"]},
-        follow=True,
-    )
-    assert response.context["request"].user == admin
 
 
 def test_create_shipping_zones(db):
@@ -183,15 +167,6 @@ def test_create_gift_card(db):
     assert GiftCard.objects.count() == 1
 
 
-def test_manifest(client, site_settings):
-    response = client.get(reverse("manifest"))
-    assert response.status_code == 200
-    content = response.json()
-    assert content["name"] == site_settings.site.name
-    assert content["short_name"] == site_settings.site.name
-    assert content["description"] == site_settings.description
-
-
 @override_settings(VERSATILEIMAGEFIELD_SETTINGS={"create_images_on_demand": False})
 def test_create_thumbnails(product_with_image, settings):
     sizeset = settings.VERSATILEIMAGEFIELD_RENDITION_KEY_SETS["products"]
@@ -240,45 +215,6 @@ def test_storages_not_setting_s3_bucket_domain(storage, settings):
     assert storage.custom_domain is None
 
 
-def test_set_language_redirects_to_current_endpoint(client):
-    user_language_point = "en"
-    new_user_language = "fr"
-    new_user_language_point = "/fr/"
-    test_endpoint = "checkout:index"
-
-    # get a English translated url (.../en/...)
-    # and the expected url after we change it
-    current_url = reverse(test_endpoint)
-    expected_url = translate_url(current_url, new_user_language)
-
-    # check the received urls:
-    #   - current url is english (/en/) (default from tests.settings);
-    #   - expected url is french (/fr/)
-    assert user_language_point in current_url
-    assert user_language_point not in expected_url
-    assert new_user_language_point in expected_url
-
-    # ensure we are getting directed to english page, not anything else
-    response = client.get(reverse(test_endpoint), follow=True)
-    new_url = response.request["PATH_INFO"]
-    assert new_url == current_url
-
-    # change the user language to French,
-    # and tell the view we want to be redirected to our current page
-    set_language_url = reverse("set_language")
-    data = {"language": new_user_language, "next": current_url}
-
-    redirect_response = client.post(set_language_url, data, follow=True)
-    new_url = redirect_response.request["PATH_INFO"]
-
-    # check if we got redirected somewhere else
-    assert new_url != current_url
-
-    # now check if we got redirect the endpoint we wanted to go back
-    # in the new language (checkout:index)
-    assert expected_url == new_url
-
-
 def test_convert_weight():
     weight = Weight(kg=1)
     expected_result = Weight(g=1000)
@@ -308,13 +244,7 @@ def test_delete_sort_order_with_null_value(menu_item):
     menu_item.delete()
 
 
-def test_csrf_middleware_is_enabled():
-    csrf_client = Client(enforce_csrf_checks=True)
-    checkout_url = reverse("checkout:index")
-    response = csrf_client.post(checkout_url)
-    assert response.status_code == 403
-
-
-def test_get_country_name_by_code():
-    country_name = get_country_name_by_code("PL")
-    assert country_name == "Poland"
+def test_placeholder(settings):
+    size = 60
+    result = placeholder(size)
+    assert result == "/static/" + settings.PLACEHOLDER_IMAGES[size]