+++ tests/test_postgresql_search.py
@@ -1,12 +1,11 @@
 from decimal import Decimal
 
 import pytest
-from django.urls import reverse
 from prices import Money
 
-from saleor.account.models import Address, User
-from saleor.order.models import Order
+from saleor.account.models import Address
 from saleor.product.models import Product
+from saleor.search.backends.postgresql import search_storefront
 
 PRODUCTS = [
     ("Arabica Coffee", "The best grains in galactic"),
@@ -31,10 +30,9 @@ def named_products(category, product_type):
     return [gen_product(name, desc) for name, desc in PRODUCTS]
 
 
-def search_storefront(client, phrase):
-    """Execute storefront search on client matching phrase."""
-    resp = client.get(reverse("search:search"), {"q": phrase})
-    return [prod for prod, _ in resp.context["results"].object_list]
+def execute_search(phrase):
+    """Execute storefront search."""
+    return search_storefront(phrase)
 
 
 @pytest.mark.parametrize(
@@ -51,10 +49,8 @@ def search_storefront(client, phrase):
 )
 @pytest.mark.integration
 @pytest.mark.django_db
-def test_storefront_product_fuzzy_name_search(
-    client, named_products, phrase, product_num
-):
-    results = search_storefront(client, phrase)
+def test_storefront_product_fuzzy_name_search(named_products, phrase, product_num):
+    results = execute_search(phrase)
     assert 1 == len(results)
     assert named_products[product_num] in results
 
@@ -67,9 +63,9 @@ def unpublish_product(product):
 
 @pytest.mark.integration
 @pytest.mark.django_db
-def test_storefront_filter_published_products(client, named_products):
+def test_storefront_filter_published_products(named_products):
     unpublish_product(named_products[0])
-    assert search_storefront(client, "Coffee") == []
+    assert not execute_search("Coffee")
 
 
 USERS = [
@@ -91,47 +87,3 @@ def gen_address_for_user(first_name, last_name):
         postal_code="53-601",
         country="PL",
     )
-
-
-@pytest.fixture
-def orders_with_addresses():
-    orders = []
-    for pk, name, lastname, email in ORDERS:
-        addr = gen_address_for_user(name, lastname)
-        user = User.objects.create(default_shipping_address=addr, email=email)
-        order = Order.objects.create(user=user, billing_address=addr, pk=pk)
-        orders.append(order)
-    return orders
-
-
-@pytest.fixture
-def orders_with_user_names():
-    orders = []
-    for pk, first_name, last_name, email in ORDERS:
-        user = User.objects.create(
-            email=email, first_name=first_name, last_name=last_name
-        )
-        order = Order.objects.create(user=user, pk=pk)
-        orders.append(order)
-    return orders
-
-
-@pytest.fixture
-def users_with_addresses():
-    users = []
-    for firstname, lastname, email in USERS:
-        addr = gen_address_for_user(firstname, lastname)
-        user = User.objects.create(default_billing_address=addr, email=email)
-        users.append(user)
-    return users
-
-
-@pytest.fixture
-def users_with_names():
-    users = []
-    for firstname, lastname, email in USERS:
-        user = User.objects.create(
-            email=email, first_name=firstname, last_name=lastname
-        )
-        users.append(user)
-    return users