+++ CHANGELOG.md
@@ -567,6 +567,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add database sequence reset to `populatedb` - #3406 by @michaljelonek
 - Get authorized amount from succeeded auth transactions - #3417 by @jxltom
 - Resort imports by `isort` - #3412 by @jxltom
+- Drop storefront 1.0 - #5043 by @IKarbowiak
 
 ### Dashboard 2.0